chore: apply api config to rest + grpc  + internal

+++ dozer-types/src/models/api_config.rs
@@ -5,29 +5,29 @@ pub struct ApiIndex {
     pub primary_key: Vec<String>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone)]
+#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone, Default)]
 pub struct ApiConfig {
     pub rest: ApiRest,
     pub grpc: ApiGrpc,
     pub auth: bool,
     pub internal: ApiInternal,
 }
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone)]
+#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone, Default)]
 pub struct ApiRest {
-    pub port: u32,
+    pub port: u16,
     pub url: String,
     pub cors: bool,
 }
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone)]
+#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone, Default)]
 pub struct ApiGrpc {
-    pub port: u32,
+    pub port: u16,
     pub url: String,
     pub cors: bool,
     pub web: bool,
 }
 
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone)]
+#[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone, Default)]
 pub struct ApiInternal {
-    pub port: u32,
+    pub port: u16,
     pub host: String,
 }
